+++ src/full/Agda/TypeChecking/Reduce.hs
@@ -1568,7 +1568,11 @@ instance InstantiateFull CompareAs where
   instantiateFull' AsTypes       = return AsTypes
 
 instance InstantiateFull Signature where
-  instantiateFull' (Sig a b c) = uncurry3 Sig <$> instantiateFull' (a, b, c)
+  instantiateFull' (Sig a b c d) = Sig
+    <$> instantiateFull' a
+    <*> instantiateFull' b
+    <*> instantiateFull' c
+    <*> pure d             -- The instance table only stores names
 
 instance InstantiateFull Section where
   instantiateFull' (Section tel) = Section <$> instantiateFull' tel
@@ -1721,6 +1725,7 @@ instantiateFullExceptForDefinitions'
     <$> ((\s r -> Sig { _sigSections     = s
                       , _sigDefinitions  = sig ^. sigDefinitions
                       , _sigRewriteRules = r
+                      , _sigInstances    = sig ^. sigInstances
                       })
          <$> instantiateFull' (sig ^. sigSections)
          <*> instantiateFull' (sig ^. sigRewriteRules))